Update on 07/06/2019
Modified: *Add Port window now allows multiple ports to be added sequentially in the same modules that have been added. *When inserting a new module, now it request the name of it, which will be displayed inside the module. *The module have changed from black block to transparent block delimited with orange border. *While selecting, a blue squared area appears indicating the zone that is being selected. To Be Done: *Try to center modules names. *Add port names over connections and circles over yet-to-be-connected ports. *Add width of wire over it. *Implement grid-like distribution in order to make it easier for the user to organize wires (most probably, granularity of 3 or 8 pixels between each point)